Add support for container insights
We want to use container insights to get more advanced metrics from
CloudWatch. This requires adding support for the `setting` block
to define cluster settings.

Currently, `containerInsights` is the only setting supported.
The valid values for this setting are `enabled` and `disabled`.

Update documentation to add new variable and new output.